requirements.txt
ollama
pydantic
Pillow
google-genai
